requirements.txt
Pillow==9.0.0
termcolor==1.1.0
